Update Lfg_Solo.cpp
AzerothCore docker installation build procedure prevents any warning to be thrown, halting the compilation if it finds any.
Since in this snippet of code the player variable is not actually used, the compiler would throw the below error:  
/azerothcore/modules/mod-solo-lfg/src/Lfg_Solo.cpp:21:25: fatal error: unused parameter 'player' [-Wunused-parameter]
   void OnLogin(Player* player)
                        ^

To prevent it, we add __attribute__ ((unused)) before the (actually unused!) player variable. This allows normal compilation.